Rewrite to TypeScript
