+++ lib/domain/press-release-rss/types.ts
@@ -1,34 +1,16 @@
-import type { DownloadResource } from "../../infrastructures/types.ts";
+import type { DownloadResource } from "../../infrastructures/index.ts";
 import type { ServiceFactory } from "../../types.ts";
-import type { RecallPressReleaseFeedItem, RssFeedItem } from "../types.ts";
-
-export type DownloadPressReleaseRss = () => Promise<Uint8Array>;
-
-export type DownloadPressReleaseRssDependencies = {
-  downloadResource: DownloadResource;
-};
-
-export type DownloadPressReleaseRssFactory = ServiceFactory<
-  DownloadPressReleaseRss,
-  DownloadPressReleaseRssDependencies
->;
-
-export type DecodeShiftJIS = (input: Uint8Array) => string;
-
-export type ParseRss = (rss: string) => RssFeedItem[];
-
-export type PickRecallPressReleaseFeedItems = (
-  feedItems: RssFeedItem[],
-) => RecallPressReleaseFeedItem[];
+import type { RecallPressReleaseFeedItem } from "../types.ts";
 
 export type RetrieveRecallPressReleaseFeedItems = () => Promise<
   RecallPressReleaseFeedItem[]
 >;
 
-export type RetrieveRecallPressReleaseFeedItemsDependencies =
-  DownloadPressReleaseRssDependencies;
+export type RetrieveRecallPressReleaseFeedItemsDependencies = {
+  downloadResource: DownloadResource;
+};
 
 export type RetrieveRecallPressReleaseFeedItemsFactory = ServiceFactory<
   RetrieveRecallPressReleaseFeedItems,
-  DownloadPressReleaseRssDependencies
+  RetrieveRecallPressReleaseFeedItemsDependencies
 >;
