+++ lib/domain/recall-press-release-page/utils.ts
@@ -1,10 +1,11 @@
+import type jsdom from "jsdom";
+import type { PdfLink } from "../types.ts";
+
 /**
  * プレスリリースページから最初の文章を抽出します。
  * 例: ○○から、令和○年○月○日国土交通大臣に対して、下記のとおりリコールの届出がありましたので、お知らせします。
- * @param {import("jsdom").JSDOM} dom
- * @returns {string}
  */
-export function extractPreamble(dom) {
+export function extractPreamble(dom: jsdom.JSDOM): string {
   const result = dom.window.document.querySelector(".date ~ p")?.textContent;
   if (!result) throw new Error("result is empty");
   return result;
@@ -12,12 +13,10 @@ export function extractPreamble(dom) {
 
 /**
  * プレスリリースページからPDFのリンクを抽出します。
- * @param {import("jsdom").JSDOM} page
- * @returns {import("../types.ts").PdfLink[]}
  */
-export function extractPdfLinks(page) {
-  const anchors = /** @type {HTMLAnchorElement[]} */ ([
+export function extractPdfLinks(page: jsdom.JSDOM): PdfLink[] {
+  const anchors = [
     ...page.window.document.querySelectorAll(".linkArrow01 a[href$='.pdf']"),
-  ]);
+  ] as HTMLAnchorElement[];
   return anchors.map((x) => ({ title: x.textContent ?? "", href: x.href }));
 }
