+++ lib/infrastructures/convert-pdf-to-images/index.ts
@@ -1,7 +1,9 @@
 import { pdf } from "pdf-to-img";
+import type { ConvertPdfToImages } from "./types.ts";
 
-/** @type {import("./types.ts").ConvertPdfToImages} */
-export const convertPdfToImages = async (pdfBuffer) => {
+export type * from "./types.ts";
+
+export const convertPdfToImages: ConvertPdfToImages = async (pdfBuffer) => {
   const results = [];
   const doc = await pdf(pdfBuffer, { scale: 1.5 });
   for await (const page of doc) {
