--- lib/infrastructures/types.ts
@@ -1,7 +1,0 @@
-export type * from "./ask-ai-to-choose-tool/types.ts";
-export type * from "./convert-pdf-to-images/types.ts";
-export type * from "./download-resource/types.ts";
-export type * from "./extract-tables-from-pdf/types.ts";
-export type * from "./posted-urls-repository/types.ts";
-export type * from "./report-error/types.ts";
-export type * from "./toot-service/types.ts";
