+++ package.json
@@ -11,6 +11,7 @@
   "license": "MIT",
   "private": true,
   "scripts": {
+    "build": "esbuild index.ts --bundle --outfile=index.mjs --platform=node --format=esm --packages=external",
     "fix": "prettier --ignore-path .gitignore --write . !package-lock.json !tests/fixtures/**",
     "test": "tsc -noEmit && tsx --test $(find tests -name '*.test.ts')"
   },
