+++ tsconfig.json
@@ -9,5 +9,5 @@
     "noFallthroughCasesInSwitch": true,
     "noEmit": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "index.mjs"]
 }
